archive/defunct_20250926_090506/mlb_backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
fastapi==0.115.12
h11==0.16.0
httpx>=0.27
idna==3.10
joblib==1.4.2
numpy==2.2.5
pandas==2.2.3
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests>=2.31.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
threadpoolctl==3.6.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
supabase
python-dotenv==1.1.0
PyYAML==6.0.2
MLB-StatsAPI==1.9.0
psycopg[binary]>=3.1

postgrest>=0.14
